backend/requirements.txt
# --- Your existing ones ---
absl-py==2.3.1
astunparse==1.6.3
certifi==2025.8.3
charset-normalizer==3.4.2
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
keras==3.11.1
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
numpy==2.1.3
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
protobuf==5.29.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
rich==14.1.0
six==1.17.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
pyserial==3.5
pytest==8.3.3

# --- New ones for backend ---
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.36
motor==3.6.0
pydantic==2.9.2
python-dotenv==1.0.1
sqlmodel==1.0.1
python-dotenv
slowapi==0.1.9
